requirements.txt
Django==3.2.16
djangorestframework==3.14.0
psycopg2-binary==2.9.9
djoser==2.2.2